+++ examples/rixs.py
@@ -4,9 +4,9 @@ Compute RIXS amplitudes according to Eq. (1) in 10.1021/acs.jctc.7b00636.
 
 import adcc
 from pyscf import gto, scf
-from scipy.constants import physical_constants
 
 from responsefun.evaluate_property import evaluate_property_isr
+from responsefun.misc import ev2au
 from responsefun.SumOverStates import TransitionMoment
 from responsefun.symbols_and_labels import (
     O,
@@ -20,10 +20,6 @@ from responsefun.symbols_and_labels import (
     w_n,
     w_prime,
 )
-
-
-def ev2au(ev):
-    return ev / physical_constants["hartree-electron volt relationship"][0]
 
 
 # run SCF in PySCF
